+++ CHANGELOG.md
@@ -2,10 +2,13 @@
 * CopyDbCommand - copy database command for test
 * documentation about **test** and **_secret.yml**
 
+[3.2.0]
+* functions as static methods
+* update user crontab, update nginx.conf in UpdateConfig class
+
 [2.0.2]
 * move all public functions to SymDep class
 * create Shell class, which provide common shell commands
 * update documentation
 * create specific recipe for **test** and **prod** deploy
 * update **deploy.php.example**
-*